Reindex any child organisations on save/destroy of parent
The OrganisationReslugger tests were failing in the previous
commit, because a reindex was being invoked twice instead of once.
Digging into this, we can see there is special logic in the
OrganisationReslugger to manually update the search index of
parent and child organisations after an organisation has its
slug updated.

As per the previous commit, we're now automatically doing that
for _parent_ organisations anyway - so the simple fix would be
to remove the OrganisationReslugger `parent_organisations` search
index calls. But then the method name `update_child_and_parent_organisations_in_search`
wouldn't make sense anymore, as we'd be left with just manually
updating child orgs, not parent orgs.

Rather than treat 'updating parent orgs' and 'updating child orgs'
differently, it surely makes more sense to update child orgs in
the same way as we update parent orgs, i.e. via the callback
introduced in the previous commit. This means we can delete
that extra code from the OrganisationReslugger.

I've added a test to capture reindexing of child orgs on parent
org updates. For completion, I also tried adding a test to capture reindexing
child orgs on parent org destroy, but the test fails with a
validation error (it's not possible to destroy a parent org which
has any child orgs), which makes total sense if you think about it.
So just the one additional test is fine.