Make 'dangerous' links unclickable in the Link Check Report UI
For broken and warning links, we create a clickable link in the
sidebar so that editors can click on the links and reproduce any
of the reported issues.

For 'dangerous' links that may have been taken over by a malicious
third party and could be distributing malware etc, we should take
a more cautious approach and just render the link as text, so that
editors don't accidentally click on a dangerous link.

Before:
> ![clickable link](https://github.com/user-attachments/assets/1c538365-30a8-41bc-8a32-8e5a8b97e546)

After:
> ![plain text](https://github.com/user-attachments/assets/724886c9-ece9-432f-a9c0-0e08a65805c1)

Trello: https://trello.com/c/PLcGoXXf