add Content Modelling team as code owners of Content Block Manager
This should ensure that the team gets tagged in
any PRs for the engine, and provides some more
documentation for any devs looking to understand
the code.

The team is setup here:
https://github.com/orgs/alphagov/teams/gov-uk-content-modelling

More info about the CODEOWNERS file:
https://docs.github.com/en/repositories/managing-your-repositorys-settings-and-features/customizing-your-repository/about-code-owners

As there is not a Whitehall specific Github
team/it's clear the whitehall team own the rest of
the code I don't think we need to specify any
other area of the repo.