Add `Ga4EcommerceTracker` to search pages
Implement new `ga4-search-results-setup.js` and add to `application.js`.

`ga4-search-results-setup.js` runs
`window.GOVUK.analyticsGa4.Ga4EcommerceTracker.init()`
which fires a `search_results` event if expected data attributes are
present on a page.

If the page has tabs, then the search results could be in a tab and not
visible when the user views the page. `ga4-search-results-setup.js` ensures
that the `search_results` event only fires if search results are
visible.