Ensure all HTML attachments belonging to foreign language only documents have correct locale
HTML attachments belonging to foreign language only documents should have a locale that matches the document's primary locale. Otherwise the default to English, which is a bad experience for users because the page furniture on GOV.UK will appear in English whilst the rest of the attachment is in the parent document's primary locale.