Use details/summary element for all 'bad links'
This is the design currently used for the "link checker API" broken
links section of the sidebar, so users are already familiar with it
and it provides a nice way of hiding away the detail until needed.

Before:
![before](https://github.com/user-attachments/assets/7a9a2eb5-35a5-44cc-8d6e-8d2ff59264b7)

After (closed):
![after](https://github.com/user-attachments/assets/6870b0dd-ce3d-42e0-9660-74da15fff8b7)

After (open):
![expanded](https://github.com/user-attachments/assets/5c856a22-19b6-4254-91c9-8569e41245be)

A CSS tweak was required to ensure that long links wrap nicely
in the sidebar, as per the broken links section.

Also removes the "don't show duplicate fixes" logic, as this same
logic isn't applied to the Link Checker API equivalents and there's
no clear user need for hiding the fact that multiple links need
amending.

Unifying the UX for both 'bad links' sections paves the way
forward for consolidating these checks into Link Checker API as
a next step.

Trello: https://trello.com/c/RtFaFX2h/45-whitehall-edition-publishing-and-scheduled-publishing-error-messages-in-the-sidebar