Remove 'Get Involved' and 'Take Part' pages
There are now no more 'Get involved' pages, including 'Take part'
pages. The last 'Take part' page has just been deleted as of last
week, and we're not planning on using that format again.

I've manually checked that all of the deleted files do not pull in
modules that were ONLY defined for the now deleted files - so I'm
confident this is the only code we can delete.

In the next PR, we'll be able to drop the `take_part_pages`
table.

Trello: https://trello.com/c/HwMcJkmn