Skip review_links when there are no host items
This checks if there are no host content items, and
if there aren’t, we skip the step completely. Additionally,
we have to check if the referrer is the previous step
(i.e. the back button has been clicked). If that’s the
case, then we redirect to the previous step.