Change link to host if previewing
This shows the preview at the end with a defined
call to action when previewing, to make it clearer
what is going on.

I’ve decided to remove the `is_preview` argument and use
inheritence to add behaviour for the preview to keep things
tider and remove too many conditionals.