Move calculation of value to the FormComponent
This removes the setting of the default value from the base component
to the form component. Before, if an embedded object contains a field
with the same name as the parent field and the embedded object doesn’t
have a value, then the value would be set to the value of the parent
field, which we don’t want.