Fix draft creation to check for clone attachment validity
Previously we seem to be checking for existing attachment save! as opposed to the newly cloned attachment for the new draft. Whilst debugging a ZenDesk ticket: https://govuk.zendesk.com/agent/tickets/6122813 we find that original attachment can return valid, but the newly cloned attachment fails validation and does not save. The save fails silently and the attachment is removed from the edition.

In doing this, some integration tests are failing because editions are `built` in the tests, not `created`. With the previous code, since we are saving the original edition attachment, that saves the edition to the database and commits the document. With this no longer there, document does not exist for the presenter to read content_id out of, hence the test fail. We need to create the editions for tests to run.