Fix `HtmlAttachment` reslug rake task
`HtmlAttachment` reslug rake task will now reslug. As the slug is
generated by the title of the `HtmlAttachment`, it now does not
accept `new_attachment_slug` as a parameter. Instead it will reslug
to the title of the `HtmlAttachment` (which can be changed in
whitehall).

In the event that a deleted `HtmlAttachment` with the same slug exists,
this rake task will now hard delete that attachment so that the
to be reslugged attachment can assume the slug.