remove `email_address` as top level block
as with postal address, this won't be a type we
create on its own.

This was the first type of block that we created
when building the CBM, so it is used in hundreds
of places - where relevant I've replaced it with
`pension`, as that is the only type in use.

I've also split out some feature tests into more
generic categories - validate and draft.