Remove `analytics-ga4.js` from `application.js`
`analytics-ga4.js` is included in `load-analytics.js`, so we were
including the same JS twice.

For tests to work, we need the modules that `analytics-ga4.js` adds
to `window`. Therefore this commit adds `test-dependencies.js` which
contains `analytics-ga4.js`. This is added to tests by updating the
`manifest.js` (which adds `test-dependencies.js` to the `public` folder)
.`jasmine-browser.json` has also been updated to pull in
`test-dependencies.js` in the tests.