Add `data-ga4-ecommerce-start-index`
For the containers that already have `data-ga4-ecommerce` data
attributes we also need to add: data-ga4-ecommerce-start-index,
the offset of the search results using the current page and the number
of results per page.

As part of this, a new helper function `page_start_index` has been added
which uses the `page` parameter to calculate this attribute.