Add feature flag for flexible pages
Working out how to add a test for the helper proved trickier than expected, because the helper depended on the `can?` method from the base application controller for authorisation, which couldn't be accessed from the test.

We decided to use allow the various edition model classes to instantiate their own enforcers, an approach for which there was an existing, albeit private method.

We also exclude flexible pages from the republishing UI for now. We will add them once the feature is stable in production.