Refactor the new document controller
The new document controller was overreliant on a complex helper class which had taken on responsibilities beyond what a Rails view helper should have, such as applying an authorisation scope to a list of classes.

We therefore refactored some of the helper behaviour, such as the feature flag check and the authorisation check, on to the document model. However, because the behaviour was very specific to the new document page, we decided that a view model for that page was the most appropriate solution. We did consider a view component, but felt that testing this logic via the DOM again did not make sense.

The refactor has allowed us to safely remove some of the tests from the new document controller, becausewe have a test on the view model which confirms that authorisation has been performed. We have unit tests in the authorisation layer which check that authorisation of fatality notice creation is enforced correctly. We can therefore safely remove those tests from the controller.