Hide attributes if object is embeddable as a block
If an object is embeddable as a block, we want to show this front and
centre, but also give the user the opportunity to use the individual
attributes that make up the block if they want.

If an object has a block-level representation, we render a `details`
component with a further summary list inside. The user can then expand
the details element to see the other blocks.

I’ve altered the styling a bit to make the component look like one
unified component, whilst still keeping the HTML neat