Add an endpoint to select a subschema for a group
This makes the `subschema` param optional in the embedded_objects#new
endpoint. If the param is not present, we fetch all the subschemas for
a group and present a list of possible schemas to the user to select.
In the next commit, we’ll add an endpoint to redirect the user to the
appropriate schema.