Remove `button_to_generates_button_tag = false`
`button_to_generates_button_tag` in the config overwrites the default
functionality of `button_to` as of Rails 7. It results in `button_to`
rendering either an `input` or `button` depending on how it is called.

The config flag was added after the Rails 7 upgrade to prevent breaking
tests but the tests breaking are not being used as of the Design System
migration, so they can be removed and we can have the intended and
consistent behaviour of `button_to`.