Add an object component
This acts in a similar way to the form components, fetching components
for each field and rendering them inside a fieldset. I tried to unit
test the component in a similar way to the form components, but this
proved tricky as they’re being rendered inside a block, so went with
the more traditional viewcomponent testing approach.