Make `GovspeakLinkValidator` a standard ActiveRecord validator
This validation class living in the DataHygiene module was a bit
of a smell, and led me up the garden path in #10279 where I ended
up implementing an additional validator in the DataHygiene module
and going down the rabbit hole of having to reference it in both
EditionPublisher and EditionScheduler.

The proper approach, achieved in this commit, is to encapsulate
the model validation inside the model (i.e. move GovSpeak link
validation into an ActiveRecord validator referenced by the
Edition model) and strip back EditionPublisher/EditionScheduler
to just defer to the `edition.valid?` logic (as they already do),
and for all other validation within those classes to be state
specific. In other words, EditionPublisher checks whether the
state of the edition (draft, published, etc) means it is allowed
to transition to the state of "published"; similar in EditionScheduler.
Edition-level concerns such as "does it contain broken links" are
all encompassed in the `edition.valid?` checks as they should be.

This does lead to some very modest UX differences, outlined below:

| Before (version on Integration) | After (local version) |
|---|---|
|![](https://github.com/user-attachments/assets/8943d9f6-3221-49c7-9907-73dc04c39fd4)|![](https://github.com/user-attachments/assets/8d8d6dc8-3da9-405d-8252-9b49a83a05ad)|
|No button to Force Publish - but repetition of broken links guidance|Button to force publish, but then a top-level error message explaining why publishing isn't possible|

This is arguably a better UX, and in any case is worth the
simplification of the code. It also means we'll be able to archive
the [backlog item](https://trello.com/c/cdDTNOQq/895-consolidate-the-two-bad-links-areas-of-the-sidebar)
we had around consolidating the "GovSpeak link check" section with
the "Link Checker API report" section, since we're now removing
the former altogether.

With this refactor merged, I'll be free to continue with #10279,
this time implementing the GovSpeak/Contact embed checks in a
standard ActiveRecord validator referenced by the Edition model
as we've done here.

Trello: https://trello.com/c/Jt5qrB9D