Add an ArrayComponent to support arrays
This adds an ArrayComponent (and accompanying ArrayItemComponent)
to support attributes that expect arrays using the `add_another`
component.

If there are no items present, we present one blank ArrayItemComponent
(which can be an fieldset containing multiple fields, or an input
field), as well as sending a component to the `empty` argument, which
is used to generate new fields when the “Add new” button is pressed.

If there are items present, we loop through each item, and generate an
ArrayItemComponent for each item, with the values prepopulated, as well
as sending an empty component to the `empty` argument.