Render inputs for flexible page content
We are using JSON schema to configure the flexible page types, which allows us to easily validate the content against the schema.

We are currently only handling object and string inputs, we can add additional input types as we add more flexible page blocks. Objects automatically wrap their child inputs in a fieldset.

The helper uses recursion to walk the JSON schema tree and build the HTML form for the flexible content. We use a dot-separated path variable to allow the helper to locate itself in the stack. JSON schema references are built using forward slashes, so we might consider whether that could be a more ergonomic choice in future.