Allow users to select a flexible page type and create an empty flexible page
We had to introduce a way to override the new document selection process because flexible pages have an extra type selection step before the user reaches the new edition form. We therefore added the `choose_document_type_form_action` to the edition model, which is overridden for flexible pages to point to the 'choose_type' action on the flexible page controller.

We have assumed that flexible pages do not need the previously published input available for now, but it may be required later.

We decided to load all the flexible page type configuration at application startup and persist it in memory. This allows us to provide a more ergonomic API for interacting with flexible page type definitions and allows the app to work without having to be aware of the file names on disk to load the type information.