Add Publishing API settings for flexible page type configuration
This commit adds a settings section so that we can control the payload sent to Publishing API for the flexible page type. We do not know what the preferred rendering app, base path, schema name or document type are going to be for flexible pages, so it is useful to be able to configure them. This allows us to leverage existing schemas, for example the history schema for the history pages. This will be important if we decide we would like to migrate existing Whitehall content to flexible pages at a later date.