Allow nested arrays to be validated
This requires us to make the `compact_nested` method work for more than
just nested hashes, as well as update `key_with_optional_prefix` to
work with nested objects too. I’ve made the latter method public, so we
can test it and document the functionality a bit better too.

I also had to fix a test that was checking for an empty string to be
`nil`, as any empty items are removed by the validator via
`compact_nested`