Add "View {current state} edition" action to editionable content
What
---

Adds a "View {current state} edition" button in the sidebar, so that:

1. We can see the contents of a published Edition without creating a new edition(!).
2. We can easily see what the contents of superseded editions used to be.

The view reuses the "Edit" view, but with the publish/settings actions removed and with an overall `<fieldset>` wrapping the form inputs to force every input to be disabled / readonly. NB, we don't add the button to the sidebar of _draft_ editions, since the publisher would just press the "Edit draft" button. Note also that we've had to fix up some tests that were referring to the wrong path and relying on the redirect from `/edit` to the parent path: now that we're allowed to view a (readonly version of a) non-draft edition on the Edit screen, there's no longer a redirect and error message.

TBC: Notably absent is the ability to view HTML attachments (or indeed any attachments) from this screen, but I figure that's something that can be added later in response to publisher demand.

Trello: https://trello.com/c/wCnq2kT4/3836-add-view-current-edition-feature

Why
---

Not being able to see an edition's GovSpeak contents impedes our ability to troubleshoot when responding to support requests. We have to go through the somewhat arduous process of switching to integration, so that we can take the non-idempotent action of creating a new edition.

Even worse is the experience of trying to find out the GovSpeak contents of past editions, which either means trying to construct the correct [diff url](https://github.com/alphagov/whitehall/blob/3aa1268058476ddfd04b135cc53cd342db51cdec/config/routes.rb#L210) to compare the superseded edition with the latest edition, or else fire up a Rails console to dig out the details manually. There is surely a tonne of benefit to more easily surfacing this information for publishers.

Screenshots
---

Sidebar on superseded editions:
> ![sidebar on superseded editions](https://github.com/user-attachments/assets/825dae06-f7f7-4014-b4e3-9f2a98e004a4)

Sidebar on 'current' editions:
> ![sidebar on 'current' editions](https://github.com/user-attachments/assets/d51bcab8-f2ee-4ceb-8b02-6ee4da3eb573)

Sidebar on draft/submitted/rejected editions (unchanged):
> ![sidebar on draft editions](https://github.com/user-attachments/assets/db7a066c-394a-4851-8df1-2946a4044612)

Read-only form:
> ![read-only form)](https://github.com/user-attachments/assets/b22858b6-aea5-479e-bd63-8bcdf92d3241)