Refactor Cucumber steps
Setting up schemas in data tables is a bit too verbose and confusing
as we’re passing round data objects and restructuring them. It’s easier
to just provide the JSON as text blocks and create the schemas that way,
especially as we’ll be adding yet more complexity in the next commit.