Add support for govspeak to flexible pages
We use JSON schema's format keyword to indicate that a string property is formatted as govspeak. Govspeak attributes are rendered as a textarea instead of an input on the edition editing form.

Leveraging our existing payload builder pattern to render the govspeak to HTML when presenting to Publishing API offers us the opportunity to decouple the flexible content presentation from the rest of the presenter. I think this will be beneficial because the presentation of different block formats is likely to become somewhat complex as we add more of them.
Co-authored-by: Alex Newton <alex.newton@digital.cabinet-office.gov.uk>
Co-authored-by: Ryan Brown <137083285+ryanb-gds@users.noreply.github.com>