Add group steps to the workflow steps
If subschemas belong to a group, we embed that group in the list
of steps in the workflow. In a similar way to how subschemas are
skipped if we’re editing a block and there are no entries for that
schema, we also skip a group step if there are no entries for schemas
in that group.