Change GovSpeak Contact Validator to check HTML Attachment validation from parent edition
Before this change, the validator in triggered on the parent edition only on publish, however the validator on HTML attachment is triggered all the time, as there is no direct publish method on the HTML attachment. This means that validation happens all the time on HTML attachments. This is causing problems for publishers as when they are trying to create a new edition but the HTML attachment contains invalid GoVSpeak, it causes validation to fail on the parent edition and no new edition is allowed to be created.
I'm solving this by getting the validator to check html attachments from the parent edition at the time of publish for the parent edition. We can't hook the validator for HTML attachment on anything as publishing for the attachments are triggered from Publishing Workeres, enqueued from pulishing parent editions.