Move `govspeak_headers` into `HeadingHierarchyValidator`
Again, this was only being used (indirectly) by DetailedGuidance,
in one place.

This also meant that the default of only pulling out h2s was
useless code, as we were always specifying the heading levels we
want to retrieve. We can therefore remove that default arg, remove
the associated test, and have confidence it still works because of
the govspeak_header_hierarchy test from the previous commit.