Give Corporate Information Pages the 'View {state} edition' feature
This was trickier than the other Edition types as it exists as
a sub resource within Organisations, and thus the generated 'view'
path needed customisation, as well as the corresponding view
itself (which wasn't making use of the `standard_edition_form`
helper method).

Corporate Information Pages _used_ to use that method, but this
was changed out in https://github.com/alphagov/whitehall/pull/8637
in order to update the 'Cancel' link destination, and with an
acknowledgement in the PR comments that they'd like to revisit and
refactor later.

I've now changed this back to `standard_edition_form` and given
the `save_or_continue_to_cancel` template the responsibility of
deciding its link path. This doesn't seem too controversial as
there is already some existing logic for determining the link
path: we're just adding a little more logic to it. That seems a
reasonable trade-off given we're then able to make use of the
standard_edition_form helper and no longer have to worry about
keeping this form in sync with the more standard Edition types
(it had already fallen behind in terms of
`ga4-visual-editor-event-handlers` etc).

This commit also adds some test for the `view_path_for_edition`
helper method I added earlier in the PR, now that it's doing a
bit more than just returning nil for GenericEdition.