Add image select format flexible page blocks
The image select block allows the user to select an image uploaded from the images tab for the edition. We add a new setting called "images_enabled" to the flexible page configuration which controls whether or not images can be uploaded.

Annoyingly, the image ID is treated by Rails as a String because, as the image value is part of a JSON attribute, we don't benefit from Rails' general ability to infer the correct type of the request parameter from the underlying database column type. This means we have to call to_i on it before we can use it to check for a matching Image ID.

At present we are not validating that the image from the request object is actually associated with the flexible page, because there isn't currently an elegant way to make the edition available to the schema validator. This does potentially allow users to "hack" the system by sending a manipulated request with a different image ID. Long term, it would be preferable to separate content validation from schema validation anyway.