Add a custom opening hours component
This adds a custom component for opening hours, which allows us to
create 0-many opening hours. This is hidden behind a checkbox, which
shows the component if it is checked.

Because the designs show drop down options for the days and times, we
need to specify select boxes for the days, as well as seperate select
boxes for each part of the time. Times are provided as a single string,
so we also need to seperate those out into their distinct parts to
present when editing.