Factor out all behaviour related to each block type into separate classes
This refactor avoids us having behaviour associated with each block distributed across the codebase. This will make it easier to reason about the behaviour of each block type and make it easier to add new blocks when required. Each block is responsible for validation, rendering its form controls, and building its publishing API payload.

In general, we want blocks to have knowledge only of the part of the content tree and the schema that they are responsible for. However, in some cases, such as the image select block, blocks will need access to other data. To solve this problem, we create a context module that is initialised for all flexible page routes in the controller, allowing blocks to access the page should they need to. The context is also used to make the renderer from the controller available to the blocks, so that the rendering behaviour can be kept in the block itself rather than in a separate helper or erb template.

Blocks are instantiated using the factory pattern. Each block has a type and a format which corresponds to the type and format attributes in the JSON schema. Format is not a required attribute in a JSON schema, so we provide default blocks for each type.