Have Corporate Information Pages use the standard_edition_form
Corporate Information Pages _used_ to use this method, but this
was changed out in #8637 in order to update the 'Cancel' link
destination, and with an acknowledgement in the PR comments that
they'd like to revisit and
refactor later.

I've now changed this back to `standard_edition_form` and given
the `save_or_continue_to_cancel` template the responsibility of
deciding its link path. This doesn't seem too controversial as
there is already some existing logic for determining the link
path: we're just adding a little more logic to it. That seems a
reasonable trade-off given we're then able to make use of the
standard_edition_form helper and no longer have to worry about
keeping this form in sync with the more standard Edition types
(it had already fallen behind in terms of
`ga4-visual-editor-event-handlers` etc).

This will enable Corporate Information Pages to take advantage of
the feature we're introducing in the next commit.