Hide navigation tabs and show Back link on the read-only form
Also adds a test for this functionality, automatically surfaced on
all standard Edition types, e.g. via `should_allow_editing_of :publication`.
Could have defined a separate method like
`should_allow_viewing_of_non_editable :publication` but we'd have
to touch a lot of files, and this view feature is arguably an
extension of the edit feature anyway (and indeed, uses the same
route and view), so it felt fine to include this in the existing
test.