Add a preprocessor to handle opening hours
This introduces the concept of a preprocessor for different object
types, which will allow us to alter what is saved if the forms fields
have a different data shape to what Publishing API expects.

Because our opening hours time fields will have multiple select boxes,
we need to consume each select box and join them together to make a
single field. This will then be validated using a regex from the schema,
so any missing fields will result in an invalid time.

Similarly, if the `hours_available` checkbox is left unchecked, we
should strip any submitted opening hours, as we don’t want these to
be saved.