Add temporary 'ContactMigrator' service
This will be used by a data migration task in the next few commits,
to find-and-replace old/bad Contact IDs with recognised alternatives.

We're starting here with supporting standard Editions, but will
add support for HTML Attachments in the next commit.

I decided to use `Minitest::Spec` test style so that we can neatly
group the different areas of the test with `describe` blocks. We
tend to prefer RSpec over Minitest, and this is a nudge in that
direction:
https://docs.publishing.service.gov.uk/manual/conventions-for-rails-applications.html#testing-utilities

Have also consciously added a "avoids creating a new Edition in
the first place if no changes are required" test so that we can
re-run the migration if it fails part-way, and it'll skip over any
previously updated ones, in theory.