Use templates to render flexible page content blocks
The previous rendering solution for content blocks risked introducing XSS attack vectors through its use of html_safe. This solution makes use of erb templates to render the content blocks, which escape all content by default. It also saves us from passing the view context around in the block context, which was a bit of a hack anyway.

Had to rename the create method on Context, as it was triggering the Rubocop SaveBang correction (and it's quite a few tests to ignore the rule for).