Create 'FindAndReplaceWorker'
This will be used as the basis for the contacts migration script,
and will also be used to simplify the RemoveDangerousLinksWorker.

There was a lot of crossover between the new requirement (to find
and replace arbitrary `[Contact:x]` shortcodes) and the existing
feature in RemoveDangerousLinksWorker (to find and replace
arbitrary dangerous URLs). Rather than reinvent the wheel, we're
now extracting one commong "FindAndReplaceWorker" which can be
called in both cases. See next commits.

Note that Sidekiq doesn't support keyword arguments
(https://github.com/sidekiq/sidekiq/wiki/Best-Practices#1-make-your-job-parameters-small-and-simple)
so we're having to pass a hash with string keys.