Add tests for attribute content rendering
There is an issue with the rendering of nested attributes, which will be fixed in a subsequent refactor.

After a first `DefaultObject` render pass, the content (previous `safe_content`) is sanitized and looks like `"{\"nested_attribute\"=&gt;\"foo\"}"`, so when we try to extract the content again, on the nested attribute's render execution, `content[property_key]` returns "nested_attribute" rather than "foo".