Module to set default label when selecting a type
Most of the time, the label will be the same as the type, so let’s help
the user and set the automatically.

There’s some added complication when a user clicks to create a new
telephone number, so we need to wait for the “Add another” button to
be visible. When it is clicked, we reinitialize the module, so all the
form elements are picked.