Set font size specifically in content
When we added the `.govspeak` class in https://github.com/alphagov/whitehall/pull/10349
this made the values render strangely, as the font size is set to
`1rem`, which makes the text render slightly smaller than default.

Setting this to `1em` resets the font size to match the rest of the
copy on the page.