Make FindAndReplaceWorker update body of HTML attachments
We want to find-and-replace Contact embed codes in HTML attachments
as well as in the parent Edition bodies. As a happy side-effect
of consolidating the "workflow logic" from the
RemoveDangerousLinksWorker, we're also now getting (for free) the
ability to auto-remove dangerous links from HTML Attachments, which
were previously ignored.

Note that whenever a new draft edition is created, the HTML
attachments for the previous edition are _cloned_ and given new
IDs: hence the slightly gnarly slug-based HTML attachment
identification.