Add 'log_prefix' keyword arg
This will be useful for picking out the results of a bulk find and
replace. E.g. in the next commit(s) where we bulk-replace Contact
embed shortcodes: having a log prefix allows us to search the logs
just for that prefix, to retrieve all logs related to the change.