Add a `_destroy` param to permitted array params
This will allow us to work out if an item should be removed before
saving.

Out of the box, the [add another component](https://components.publishing.service.gov.uk/component-guide/add_another)
provides functionality for a “destroy” checkbox, which is checked when
the user hits delete (and provides a JS-less fallback). We have these
checkboxes at the moment, but they don’t work. We’ll fix this in a
future commit.