Only render checkbox on mutable items
This checks if an item with a given index is present in the currently
published version of an object. If it is, we render a hidden field with
a `_destroy` value of `0`. If not, we render a hidden checkbox which,
if checked, has a `_destroy` value of `1`.

When the user hits `Delete` on that item, as well as hiding the item, it
also checks the hidden checkbox, allowing us to delete the item using
the functionality we added previously.